Refactored YahtzeeScoring
The intention of this commit is to refactor the code in YahtzeeScoring while maintaining compatibility with tests.

There were a few reasons why I decided to refactor. First, I realized that there was no case in which the Upper Section would score higher than the Lower Section. Since the Upper Section never considers the sum of the entire roll, then it follows that any combination of dice will result in a higher sum (as a whole roll) or a specific pattern (such as a straight or a full house).

Next, the original implementation didn't follow a very OOP architecture. For example, the `roll` variable was being passed to every method signature. Before adding an initializer, I was having to pass `roll` through the call stack while trying to refactor code. Adding the initializer made the class able to rely on the instance variable `@roll`, which improved readability and encapsulated the data of the class in one convenient place.

Finally, the original implementation had some instances of repetition in it. For example, `#score_three/four_of_a_kind` & `#score_small/large_straight` follow similar logic between methods. The logic was also missing an edge case where certain combinations of rolls that appear to be full houses would be better scored as a three of a kind.